requirements.txt
# Agent Tools
pydantic
beautifulsoup4
requests
duckduckgo_search

# Ell
ell-ai

# Other
python-dotenv